+++ CHANGELOG.md
@@ -2,11 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.0-beta.44] - 2020-06-11
+
+### Added
+
+-   add `DVDPrime` search
+
 ## [1.0.0-beta.42] - 2020-06-11
 
 ### Added
 
--   add `dcinside` search (`w`)
+-   add `dcinside` search
 
 ## [1.0.0-beta.41] - 2020-06-10
 
@@ -36,6 +42,7 @@ All notable changes to this project will be documented in this file.
 -   fix auto refresh bug
 -   update [README.md](README.md)
 
+[1.0.0-beta.44]: https://github.com/samnoh/cliboards/releases/tag/v1.0.0-beta.44
 [1.0.0-beta.42]: https://github.com/samnoh/cliboards/releases/tag/v1.0.0-beta.42
 [1.0.0-beta.41]: https://github.com/samnoh/cliboards/releases/tag/v1.0.0-beta.41
 [1.0.0-beta.40]: https://github.com/samnoh/cliboards/releases/tag/v1.0.0-beta.40